requirements.txt
aiohttp==3.7.4.post0
aiomysql==0.0.21
aiopg==1.3.2
aiosignal==1.2.0
aiosqlite==0.17.0
async-timeout==3.0.1
attrs==21.2.0
bottle==0.12.19
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
discord-components==2.1.2
discord.py==1.7.3
discordSuperUtils==0.2.8
ffmpeg-python==0.2.0
frozenlist==1.3.0
future==0.18.2
greenlet==1.1.2
idna==2.8
lavalink==3.1.5
monotonic==1.6
motor==2.5.1
multidict==4.7.6
opuslib==3.0.1
Pillow==9.1.1
platformdirs==2.2.0
playsound==1.3.0
playwright==1.12.1
psycopg2-binary==2.9.1
pycparser==2.21
pyee==9.0.4
pymongo==3.12.0
PyMySQL==0.9.3
PyNaCl==1.3.0
python-dateutil==1.4
python-mpd2==3.0.5
pytz==2019.3
PyYAML==5.3.1
requests==2.22.0
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
six==1.16.0
spotipy==2.16.1
typing-extensions==3.10.0.0
urllib3==1.25.11
wavelink==0.9.10
webencodings==0.5.1
websockets==9.1
yarl==1.5.1
youtube-dl==2021.6.6
